Light warmth slider + fixed input_number slider
I have added a new configuration field called function. The default is "Brightness" but if you set it to "Warmth", it will set light warmth of light.
And fixed the input_number, before it could only do 0-100 values, but it now sets the minimum value to input_number minimum and the same with the maximum number.