Added media_player set volume support
The slider now support media_player.volume_set service. 
It was a bit tricky to make it show the correct state, because volume_level is a float from 0 - 1. So it could be 0.17333333333, which made it a bit buggy. I found a solution for this, which I am happy with. The slider is from 0 - 100. When setting the volume it calls a method that takes that integer and converts it to a float. So if the integer is 27 the volume is set to 0.27. If it is set to 100 the volume will be 1.

To use this as a volume control for your media player, simply set the entity_id as media_player.test_player.