Cover is now supported
You can now use a cover with the slider-card. Simply put that as the entity.
I am using Roller Shade, which is working fine.

I am going to add a new feature soon, where you can select vertical slider. This is because I have a use case for it now that I have smart roller shades.