BREAKING CHANGES:
I have changed the codebase for the card, this is to make development easier and more streamlined.
This has also brought the change that the card is not internet dependant anymore.
Before it had to import the LitElement library from a link, now it is compiled with directly in the code.

I have also changed the name of the card to "my-slider".
I have done this because, I am working on some new button cards, and I am planning on calling them "my-button-light" for example. So this new name fits into the theme.